Implemented the recyclerView and moved the json link to properties file, which is not commited

+++ app/src/main/java/com/tomasmichalkevic/bakingapp/data/Ingredient.java
@@ -1,3 +1,41 @@
+/*
+ * PROJECT LICENSE
+ *
+ * This project was submitted by Tomas Michalkevic as part of the Nanodegree At Udacity.
+ *
+ * As part of Udacity Honor code, your submissions must be your own work, hence
+ * submitting this project as yours will cause you to break the Udacity Honor Code
+ * and the suspension of your account.
+ *
+ * Me, the author of the project, allow you to check the code as a reference, but if
+ * you submit it, it's your own responsibility if you get expelled.
+ *
+ * Copyright (c) 2018 Tomas Michalkevic
+ *
+ * Besides the above notice, the following license applies and this license notice
+ * must be included in all works derived from this project.
+ *
+ * MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package com.tomasmichalkevic.bakingapp.data;
 
 import com.google.gson.annotations.Expose;
@@ -7,7 +45,7 @@ public class Ingredient {
 
     @SerializedName("quantity")
     @Expose
-    private int quantity;
+    private double quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -28,18 +66,18 @@ public class Ingredient {
      * @param ingredient
      * @param quantity
      */
-    public Ingredient(int quantity, String measure, String ingredient) {
+    public Ingredient(double quantity, String measure, String ingredient) {
         super();
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
